pyproject.toml
[project]
name = "chat-with-your-documents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.0",
    "fastapi[all]>=0.118.0",
    "google-genai>=1.41.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pydantic-ai>=1.0.14",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "sentence-transformers>=5.1.1",
    "tiktoken>=0.11.0",
]
